Refactor. INT in Branch is changed to uint64_t, INT in all score computations is changed to SCORE

+++ source/compareBranches.c
@@ -39,7 +39,7 @@ BranchInfo** treeToBranchInfo(Tree* tree)
     int j = 0;
     int k = 0;
     long dist = 0;
-    INT p = 0;
+    uint64_t p = 0;
     size_t notTrivialBranchesNum = 0;
     BranchInfo** branchInfoArr = NULL;
     BranchInfo** branchInfoArrRmTrivial = NULL;
